fix(stdict): add fallback view route and robust parsing for examples/synonyms/antonyms; store sense_code as int

+++ src/main/java/com/words_hanjoom/domain/wordbooks/dto/response/WordDto.java
@@ -12,7 +12,7 @@ public record WordDto(
         String wordCategory,
         Byte   shoulderNo,
         Long   targetCode,
-        Short  senseNo
+        Integer  senseNo
 ) {
     private static final Pattern STD_MARKS = Pattern.compile("[-·ㆍ‐–—]");
